refactor(dependencies): Remove circular imports
Import Iterables into maps and sets direct from source instead of via main file.
